moved configuration files to configuration folder
